feat: implement search debounce

+++ app/src/main/java/com/example/playlistmaker/network/SearchResponse.kt
@@ -8,7 +8,7 @@ data class SearchResponse(
 )
 
 data class Track(
-    val trackId: Int,
+    val trackId: Long,
     val trackName: String,
     val artistName: String,
     @SerializedName("trackTimeMillis") val trackTime: Long,
